removed old x,y,z vars and functions
commiting to direct angle control. It seems to be much more precise, and
when you tell a servo to move one degree, it actually  moves one degree.
With the x,y,z commands, sometimes the arm would not move unless you go
a certain threshold, and then it would overshoot. oh well. angles it is.