Servo package - adding cutout for cable
This allows you to push the servo through the cutout. My first design
had the servo mounted so that the top of the servo pushed through the
plexi, and this made for a pretty low profile on the gear that prodrudes
through to the other side. However, I think I need the servo to sit on
the outside of the gearbox, so that I can shift the larger gears further
out on the rods. This will allow for more clearance between the gears
and the surface of the outside of the gearbox - needed for proper end
caps.