belt tension corrections
Brought down the entire top about 1/2". The original design was too
tight for the belt to even get in there. With the bearing removed
(effectively bringing the top roller down 1/2" inch), it was able to be
placed loosely.

Added tentioning slots, where I can place a 1/2" rod with clamps and add
tension to the belt.

designed the inner wall.

Removed hinges, I don't plan on having an opening or waffle top on this.
I'd like the tech to load from top (just like the uarm will do). They
can manually move the belt one board space by pressing the "up" or
"down" buttons.